Initial Commit via file Upload
A project to determine which image classification algorithm (VGG, ALEXNET, RESNET) works the "best" on classifying images as "dogs" or "not dogs".

An image classifier is simply a tool that has an input and an output. The Input is an image. The output determines what the image depicts. (for example, a dog). 

Image classifiers do not always categorize the images correctly.

Each algorithm has been timed when solving the classification problem and accuracy reported.

I've also included comments on what the best classifier inside the text files generated by the shell script.